+++ src/components/Accounts/Actions/PayModal.js
@@ -22,16 +22,17 @@ import css from './PayWaive.css';
 
 const validate = (values, props) => {
   const accounts = props.accounts || [];
-  let selected = parseFloat(0);
+  let selected = 0;
   accounts.forEach(a => {
-    selected += parseFloat(a.remaining);
+    selected += (a.remaining * 100);
   });
+  selected /= 100;
 
   const errors = {};
   if (!values.amount) {
     errors.amount = <FormattedMessage id="ui-users.accounts.error.field" />;
   }
-  if (values.amount < 0) {
+  if (values.amount <= 0) {
     errors.amount = <FormattedMessage id="ui-users.accounts.pay.error.amount" />;
   }
   if (!values.method) {
@@ -215,6 +216,7 @@ class PayModal extends React.Component {
     const submitButtonDisabled = pristine || submitting || invalid;
     return (
       <Modal
+        id="pay-modal"
         open={open}
         label={<FormattedMessage id="ui-users.accounts.pay.payFeeFine" />}
         onClose={onClose}
@@ -231,7 +233,7 @@ class PayModal extends React.Component {
               <Row end="xs">
                 <Col xs={7}>
                   <FormattedMessage id="ui-users.accounts.pay.field.totalamount" />
-:
+                  {':'}
                 </Col>
                 <Col xs={4}>
                   {parseFloat(totalAmount).toFixed(2)}
@@ -240,7 +242,7 @@ class PayModal extends React.Component {
               <Row end="xs">
                 <Col xs={7}>
                   <FormattedMessage id="ui-users.accounts.pay.field.selectedamount" />
-:
+                  {':'}
                 </Col>
                 <Col xs={4}>
                   {parseFloat(selected).toFixed(2)}
@@ -250,7 +252,7 @@ class PayModal extends React.Component {
                 <Col xs={7}>
                   <b>
                     <FormattedMessage id="ui-users.accounts.pay.field.paymentamount" />
-*:
+                    {'*:'}
                   </b>
                 </Col>
                 <Col xs={4} className={css.customCol}>
@@ -271,7 +273,7 @@ class PayModal extends React.Component {
               <Row end="xs">
                 <Col xs={7}>
                   <FormattedMessage id="ui-users.accounts.pay.field.remainingamount" />
-:
+                  {':'}
                 </Col>
                 <Col xs={4}>
                   {remaining}
@@ -282,7 +284,7 @@ class PayModal extends React.Component {
               <Row>
                 <Col xs>
                   <FormattedMessage id="ui-users.accounts.pay.field.paymentmethod" />
-*
+                  {'*'}
                 </Col>
               </Row>
               <Row>